+++ src/items/Keys.java
@@ -93,7 +93,7 @@ public class Keys implements Interactable, Movable, Drawable {
     }
 
     @Override
-    public boolean pointNearObject(int x, int y, int z) {
+    public boolean pointNearObject(double x, double y, int z) {
         return false;
     }
 }
