+++ src/items/Player.java
@@ -99,7 +99,7 @@ public class Player implements Drawable {
     }
 
     @Override
-    public boolean pointNearObject(int x, int y, int z) {
+    public boolean pointNearObject(double x, double y, int z) {
         return false;
     }
